Removing vestigial components. Moved helper file to better place.
