--- src/components/pages/collectRecordFormPages/CollectRecordFormPage/index.js
@@ -1,3 +1,0 @@
-import CollectRecordFormPage from './CollectRecordFormPage'
-
-export default CollectRecordFormPage
